Add Longhorn Raw Block Device Support
Fio container need to be run with privileged mode to be able to invalid
the cache, and for k8s, if a container is running with privileged mode,
the volumeDevices mapping will be ignored, link is below:

https://github.com/kubernetes/kubernetes/issues/85624

We work around it by list the device name under /dev/longhorn, and pass
it to fio.

Also modified README.md to support Longhorn.

Longhorn #1664

Signed-off-by: Bo Tao <bo.tao@rancher.com>